@@ -2,4 +2,4 @@ gemspec :path => '.'
 
 instance_eval File.read(File.expand_path('../gemfiles/Gemfile.base', __FILE__))
 
-gem 'rails', '~> 3.2.0'
+gem 'rails', '~> 5.0.0'
